// be/src/common/enums/wallet-transaction-type.enum.ts
export enum WalletTransactionType {
  DEPOSIT = 'DEPOSIT',
  WITHDRAW = 'WITHDRAW',
  PAYMENT = 'PAYMENT',
  REFUND = 'REFUND',
  PLATFORM_FEE = 'PLATFORM_FEE',
  TASKER_EARNING = 'TASKER_EARNING',
  DEPOSIT_HOLD = 'DEPOSIT_HOLD',
  DEPOSIT_RELEASE = 'DEPOSIT_RELEASE',
  DEPOSIT_DEDUCT = 'DEPOSIT_DEDUCT',
  CANCELLATION_FEE = 'CANCELLATION_FEE',
  ADJUSTMENT = 'ADJUSTMENT',
}